+++ CHANGELOG.md
@@ -6,9 +6,8 @@ This changelog adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ## [1.4.0] - 2025-05-14
 
-- **Add Collapse/Expand All actions to tree views**: Action buttons added to the Regions and Full Outline tree views' title bars to collapse or expand all tree items
-  - *Bonus:* The cursor's tree item will be auto-highlighted after "Expand All" to help re-orient, regardless of the view's `shouldAutoHighlight` setting
-  - *Limitation:* The "Expand All" actions are limited to an expansion depth of 3 by VSCode's API
+- **Add Collapse/Expand All actions**: Action buttons added to the Regions and Full Outline tree views' title bars to collapse or expand all tree items
+- **Add persistent tree view state**: Expanded/collapsed items are remembered when switching files or restarting VS Code
 
 ## [1.3.0] - 2025-03-18